MintTrack/requirements.txt
asgiref==3.7.2
attrs==23.1.0
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
psycopg==3.1.10
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.5.0
